Update UI màn B và fix bug giải thuật

+++ nts.uk/uk.at/at.ctx/function/nts.uk.ctx.at.function.infra/src/main/java/nts/uk/ctx/at/function/infra/entity/processexecution/KfnctExecutionIndexCategoryPk.java
@@ -32,5 +32,5 @@ public class KfnctExecutionIndexCategoryPk implements Serializable {
 	 * 	カテゴリNO
 	 */
 	@Column(name = "CATEGORY_NO")
-	public BigDecimal categoryNo;
+	public int categoryNo;
 }
